Rename metrics2.py as metric.py
